backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
pymongo>=4.7.0
scikit-learn>=1.5.0
numpy>=2.0.0
pandas>=2.2.0
joblib>=1.4.0
python-dotenv>=1.0.0
